# openapi.yaml
openapi: 3.0.3
info:
  title: Invoice Analysis API
  description: API for uploading invoices and processing them to provide analysis.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
servers:
  - url: http://localhost:8000
paths:
  /upload:
    post:
      summary: Upload an invoice and analyze its details
      description: Uploads an invoice file and processes it to return invoice details, suggestions, and an email body.
      operationId: uploadInvoice
      tags:
        - Invoice Analysis
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                  description: The invoice file to be uploaded.
                user_type:
                  type: string
                  description: The type of user (e.g., vendor or buyer).
      responses:
        '200':
          description: Invoice processed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvoiceResponse'
        '400':
          description: Bad request
        '500':
          description: Internal server error
  /ask_chat:
    post:
      summary: Handle chat queries and return full response
      description: Accepts a chat query and document ID, retrieves relevant data, and returns a JSON response.
      operationId: askChat
      tags:
        - Chat Analysis
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                query:
                  type: string
                  description: The chat query to process.
                doc_id:
                  type: string
                  description: The document ID to retrieve relevant data.
      responses:
        '200':
          description: Chat response returned successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  response:
                    type: string
                    description: The full response from the chat engine.
        '400':
          description: Bad request
        '404':
          description: Document not found
        '500':
          description: Internal server error
  /invoices/by-email:
    get:
      summary: Retrieve invoices by user email
      description: Fetches all invoices associated with a given user email.
      operationId: getInvoicesByEmail
      tags:
        - Invoice Retrieval
      parameters:
        - name: user_email
          in: query
          required: true
          schema:
            type: string
          description: The email address of the user to retrieve invoices for.
      responses:
        '200':
          description: Successfully retrieved invoices
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Invoice'
        '404':
          description: No invoices found for this email
        '500':
          description: Internal server error
components:
  schemas:
    Invoice:
      type: object
      properties:
        invoice_number:
          type: string
        amount:
          type: number
        due_date:
          type: string
        payment_status:
          type: string
        discount_rate:
          type: number
          nullable: true
        late_fee:
          type: number
          nullable: true
        grace_period:
          type: string
        vendor_name:
          type: string
        buyer_name:
          type: string
        suggestions:
          type: array
          items:
            type: object
            properties:
              type:
                type: string
              action:
                type: string
        email_body:
          type: object
          properties:
            subject:
              type: string
            body:
              type: string
    InvoiceResponse:
      type: object
      properties:
        response:
          $ref: '#/components/schemas/Invoice'